#49: fixes subpackage naming & renames TemplatesReg struct to Registry
